feat(ui): refresh UI and add custom level option
Replace default favicon and logos with tennis-themed icons.
Update index.html title and description for SEO.
Add of a new welcome page with explanations.
Change the color palette of the app consistent branding.
Add custom level with user-defined intervals and cues.
Change the cues and the logic for the other levels.